Improve power on in low power conditions
Maiden flight failed to wake up the following morning
Testing overnight on the ground revealed that it entered a reset loop, where the MCU would crash after powering on the SI5351, likely due to inrush current. The MCU was the reset by the watchdog, so at least we know that works.

The situation was not improved with a supercap. Current theory is that the board tries to power on as soon as it gets any power, but at this point the panels can't supply enough current. When it tries to pull more current, the panel voltage collapses causing the board to reset.

This commit makes the MCU test the supercap/solar voltage. If it's above a certain threshold where there is likely to be enough current, it will continue the startup. If not, it will wait. This has worked in testing even though the weather was mostly cloudy.